UI refactor
No functionality changes
