+++ instrument/UI/Types.hs
@@ -1,0 +1,11 @@
+module UI.Types where
+
+data Name
+  = ActiveTasks
+  | TaskDetails
+  | ModuleSelector
+  | ModuleDetails
+  | SessionSelector
+  | OptionsEditor
+  | OEExtraGhcOptions
+  deriving stock (Eq, Ord, Show)
